REACT_APP_ALLOW_DELETING_FROM_DATABASE boolean now controls deleteFromDb

+++ server/src/services/deleteAccountData.ts
@@ -1,25 +1,20 @@
-import nearApi from 'near-api-js';
+import { stringToBoolean } from '../../../shared/helpers/strings.js';
 
 /* eslint-disable import/extensions */
-import getConfig from '../../../shared/config.js';
 import { TxActions } from '../models/TxActions.js';
 import { TxTasks } from '../models/TxTasks.js';
 /* eslint-enable import/extensions */
 
-const NODE_ENV = process.env.NODE_ENV || 'development';
-const nearConfig = getConfig(NODE_ENV);
-
-const { nodeUrl } = nearConfig;
-
-const connectionInfo = { url: nodeUrl };
-const provider = new nearApi.providers.JsonRpcProvider(connectionInfo);
-const connection = new nearApi.Connection(nodeUrl, provider, {});
+const ALLOW_DELETING_FROM_DATABASE = stringToBoolean(process.env.ALLOW_DELETING_FROM_DATABASE ?? 'false');
 
 // eslint-disable-next-line consistent-return
 export const deleteAccountData = async (request, response) => {
   const { accountId } = request.body;
   console.log('deleteAccountData', { accountId });
   try {
+    if (!ALLOW_DELETING_FROM_DATABASE) {
+      throw new Error('ALLOW_DELETING_FROM_DATABASE is not enabled.');
+    }
     TxTasks.deleteMany({ accountId })
       .then()
       .catch((error) => console.error({ error }));
